+++ cmd/repo-updater/internal/purge/BUILD.bazel
@@ -1,0 +1,22 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "purge",
+    srcs = ["purge.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/repo-updater/internal/purge",
+    visibility = ["//cmd/repo-updater:__subpackages__"],
+    deps = [
+        "//internal/actor",
+        "//internal/conf/conftypes",
+        "//internal/database",
+        "//internal/gitserver",
+        "//internal/goroutine",
+        "//internal/ratelimit",
+        "//lib/errors",
+        "//schema",
+        "@com_github_prometheus_client_golang//prometheus",
+        "@com_github_prometheus_client_golang//prometheus/promauto",
+        "@com_github_sourcegraph_log//:log",
+        "@org_golang_x_time//rate",
+    ],
+)
