+++ cmd/repo-updater/shared/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     deps = [
         "//cmd/frontend/envvar",
         "//cmd/frontend/globals",
+        "//cmd/repo-updater/internal/purge",
         "//cmd/repo-updater/internal/repoupdater",
         "//cmd/repo-updater/internal/scheduler",
         "//internal/actor",
@@ -41,7 +42,6 @@ go_library(
         "//internal/httpserver",
         "//internal/instrumentation",
         "//internal/observation",
-        "//internal/ratelimit",
         "//internal/repos",
         "//internal/repoupdater/v1:repoupdater",
         "//internal/service",
@@ -54,7 +54,6 @@ go_library(
         "@com_github_sourcegraph_log//:log",
         "@org_golang_google_grpc//:go_default_library",
         "@org_golang_google_grpc//reflection",
-        "@org_golang_x_time//rate",
     ],
 )
