+++ internal/batches/BUILD.bazel
@@ -14,7 +14,6 @@ go_library(
         "//internal/batches/syncer",
         "//internal/database",
         "//internal/encryption",
-        "//internal/goroutine",
         "//internal/httpcli",
         "//internal/observation",
         "@com_github_sourcegraph_log//:log",
