+++ internal/repos/BUILD.bazel
@@ -27,7 +27,6 @@ go_library(
         "pagure.go",
         "perforce.go",
         "phabricator.go",
-        "purge.go",
         "python_packages.go",
         "ruby_packages.go",
         "rust_packages.go",
@@ -49,7 +48,6 @@ go_library(
         "//internal/bytesize",
         "//internal/codeintel/dependencies",
         "//internal/conf",
-        "//internal/conf/conftypes",
         "//internal/conf/reposource",
         "//internal/database",
         "//internal/database/basestore",
